- 앱 생성시, 워크플로우 생성 - 워크플로우 조회, 수정

+++ src/main/java/kr/co/proten/llmops/api/app/service/impl/AppServiceImpl.java
@@ -1,6 +1,9 @@
 package kr.co.proten.llmops.api.app.service.impl;
 
-import kr.co.proten.llmops.api.app.dto.request.*;
+import kr.co.proten.llmops.api.app.dto.request.AppCreateDTO;
+import kr.co.proten.llmops.api.app.dto.request.AppSearchDTO;
+import kr.co.proten.llmops.api.app.dto.request.AppStateDTO;
+import kr.co.proten.llmops.api.app.dto.request.AppUpdateDTO;
 import kr.co.proten.llmops.api.app.dto.response.AppResponseDTO;
 import kr.co.proten.llmops.api.app.entity.AppEntity;
 import kr.co.proten.llmops.api.app.mapper.AppMapper;
@@ -10,6 +13,9 @@ import kr.co.proten.llmops.api.workflow.entity.WorkflowEntity;
 import kr.co.proten.llmops.api.workflow.service.WorkflowService;
 import kr.co.proten.llmops.api.workspace.entity.WorkspaceEntity;
 import kr.co.proten.llmops.api.workspace.service.WorkspaceService;
+import kr.co.proten.llmops.core.exception.InvalidInputException;
+import kr.co.proten.llmops.core.exception.StateChangeException;
+import lombok.RequiredArgsConstructor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.data.domain.PageRequest;
@@ -21,11 +27,9 @@ import org.springframework.transaction.annotation.Transactional;
 import java.util.List;
 import java.util.NoSuchElementException;
 
-import static kr.co.proten.llmops.core.helpers.DateUtil.generateCurrentTimestamp;
-import static kr.co.proten.llmops.core.helpers.UUIDGenerator.generateUUID;
-
 @Service
 @Transactional
+@RequiredArgsConstructor
 public class AppServiceImpl implements AppService {
 
     private final Logger log = LoggerFactory.getLogger(this.getClass());
@@ -34,15 +38,8 @@ public class AppServiceImpl implements AppService {
     private final WorkflowService workflowService;
     private final WorkspaceService workspaceService;
 
-    public AppServiceImpl(AppRepository appRepository, AppMapper appMapper, WorkflowService workflowService, WorkspaceService workspaceService) {
-        this.appRepository = appRepository;
-        this.appMapper = appMapper;
-        this.workflowService = workflowService;
-        this.workspaceService = workspaceService;
-    }
-
-    @Transactional
     @Override
+    @Transactional
     public AppResponseDTO createApp(AppCreateDTO appCreateDTO) {
         try {
             WorkspaceEntity workspaceDummy = createDummyWorkspace(appCreateDTO.workspace_id());
@@ -54,14 +51,10 @@ public class AppServiceImpl implements AppService {
             log.info("Workflow saved: {}", savedWF);
 
             AppEntity appEntity = AppEntity.builder()
-                    .appId(generateUUID())
                     .workspace(workspaceDummy)
-//                    .workflow(workflow)
+                    .workflow(workflow)
                     .name(appCreateDTO.name())
                     .description(appCreateDTO.description())
-                    .createdAt(generateCurrentTimestamp())
-                    .updatedAt(null)
-                    .isActive(true)
                     .build();
 
             AppResponseDTO responseDTO = appMapper.responseToDto(appRepository.save(appEntity));
@@ -70,23 +63,20 @@ public class AppServiceImpl implements AppService {
             return responseDTO;
         } catch (Exception e) {
             log.error("Error occurred during app creation: {}", e.getMessage(), e);
+            throw new InvalidInputException("앱 생성 실패!");
         }
-        return null;
     }
 
     @Override
     public AppResponseDTO getAppById(String workspaceId, String appId) {
-        try {
-            validateWorkspace(workspaceId);
-            return appMapper.responseToDto(appRepository.findById(appId)
-                    .orElseThrow(() -> new NoSuchElementException("App not found with id: " + appId)));
-        } catch (Exception e) {
-            log.error("Error retrieving app by ID: {}", e.getMessage(), e);
-            throw e;
-        }
+        validateWorkspace(workspaceId);
+
+        return appMapper.responseToDto(appRepository.findById(appId)
+                .orElseThrow(() -> new NoSuchElementException(String.format("앱 ID [%s]에 해당하는 앱을 찾을 수 없습니다.", appId))));
     }
 
     @Override
+    @Transactional(readOnly = true)
     public List<AppResponseDTO> getAppByName(AppSearchDTO appSearchDTO) {
         try {
             validateWorkspace("8ee589ef-c7bb-4f2a-a773-630abd0de8c7");
@@ -103,12 +93,12 @@ public class AppServiceImpl implements AppService {
                     .toList();
         } catch (Exception e) {
             log.error("Error searching apps by name: {}", e.getMessage(), e);
-            throw e;
+            throw new NoSuchElementException(String.format("앱 이름 [%s]에 해당하는 앱을 찾을 수 없습니다.", appSearchDTO.name()));
         }
     }
 
-    @Transactional(readOnly = true)
     @Override
+    @Transactional(readOnly = true)
     public List<AppResponseDTO> getAllApps(String workspaceId, int page, int size, String sortField, String sortBy) {
         try {
             WorkspaceEntity workspace = validateWorkspace(workspaceId);
@@ -120,33 +110,31 @@ public class AppServiceImpl implements AppService {
                     ? PageRequest.of(page, size, Sort.by(Sort.Order.asc(sortField)))
                     : PageRequest.of(page, size, Sort.by(Sort.Order.desc(sortField)));
 
-//            return appRepository.findAll(pageable).stream()
-//                    .map(appMapper::responseToDto)
-//                    .toList();
+            log.info("find list results: {}", appRepository.findAllByWorkspace(workspace ,pageable));
 
-            log.info("find list results: {}", appRepository.findAllByWorkspace(workspace));
-
-            return  null;
+            return appRepository.findAllByWorkspace(workspace ,pageable).stream()
+                    .map(appMapper::responseToDto)
+                    .toList();
         } catch (Exception e) {
             log.error("Error retrieving all apps: {}", e.getMessage(), e);
+            throw new NoSuchElementException(String.format("워크스페이스 ID [%s]에 해당하는 앱을 찾을 수 없습니다.", workspaceId));
         }
-        return List.of();
     }
 
     @Override
+    @Transactional
     public AppResponseDTO updateApp(AppUpdateDTO appUpdateDTO) {
         try {
             AppEntity existingApp = appRepository.findById(appUpdateDTO.app_id())
-                    .orElseThrow(() -> new NoSuchElementException("App not found with id: " + appUpdateDTO.app_id()));
+                    .orElseThrow(() -> new NoSuchElementException(String.format("수정할 앱 ID [%s]에 해당하는 앱을 찾을 수 없습니다.", appUpdateDTO.app_id())));
 
             existingApp.setName(appUpdateDTO.name());
             existingApp.setDescription(appUpdateDTO.description());
-            existingApp.setUpdatedAt(generateCurrentTimestamp());
 
             return appMapper.responseToDto(appRepository.save(existingApp));
         } catch (Exception e) {
             log.error("Error updating app: {}", e.getMessage(), e);
-            throw e;
+            throw new StateChangeException("업데이트 변경 실패!");
         }
     }
 
@@ -154,26 +142,24 @@ public class AppServiceImpl implements AppService {
     public AppResponseDTO updateAppState(AppStateDTO appStateDTO) {
         try {
             AppEntity existingApp = appRepository.findById(appStateDTO.app_id())
-                    .orElseThrow(() -> new NoSuchElementException("App not found with id: " + appStateDTO.app_id()));
+                    .orElseThrow(() -> new NoSuchElementException(String.format("수정할 앱 ID [%s]에 해당하는 앱을 찾을 수 없습니다.", appStateDTO.app_id())));
 
             existingApp.setActive(appStateDTO.is_active());
 
             return appMapper.responseToDto(appRepository.save(existingApp));
         } catch (Exception e) {
             log.error("Error updating app state: {}", e.getMessage(), e);
-            throw e;
+            throw new StateChangeException("상태 변경 실패!");
         }
     }
 
     @Override
     public boolean deleteApp(String id) {
-        try {
-            appRepository.deleteById(id);
-            return true;
-        } catch (Exception e) {
-            log.error("Error deleting app: {}", e.getMessage(), e);
-            throw e;
+        if (!appRepository.existsById(id)) {
+            throw new NoSuchElementException(String.format("삭제할 앱 ID [%s]에 해당하는 앱을 찾을 수 없습니다.", id));
         }
+        appRepository.deleteById(id);
+        return true;
     }
 
     @Override
@@ -182,30 +168,26 @@ public class AppServiceImpl implements AppService {
             appRepository.deleteAllById(appIdList);
             return true;
         } catch (Exception e) {
-            log.error("Error deleting app list: {}", e.getMessage(), e);
-            throw e;
+            throw new NoSuchElementException(String.format("삭제할 앱 ID [%s]에 해당하는 앱을 찾을 수 없습니다.", appIdList.toString()));
         }
     }
 
     private WorkspaceEntity validateWorkspace(String workspaceId) {
-        WorkspaceEntity workspaceDummy = createDummyWorkspace(workspaceId);
-        if (workspaceDummy == null) {
-            throw new NoSuchElementException("Workspace not found with id: " + workspaceId);
-        }
+//        WorkspaceEntity workspaceDummy = createDummyWorkspace(workspaceId);
+        WorkspaceEntity workspaceDummy = workspaceService.findWorkspaceById(workspaceId)
+                        .orElseThrow(() -> new NoSuchElementException(String.format("워크스페이스 ID [%s]에 해당하는 워크스페이스을 찾을 수 없습니다.", workspaceId)));
         return workspaceDummy;
     }
 
     private WorkspaceEntity createDummyWorkspace(String workspaceId) {
         if (workspaceId == null) return null;
-        WorkspaceEntity workspace = new WorkspaceEntity();
-        workspace.setWorkspaceId("8ee589ef-c7bb-4f2a-a773-630abd0de8c7");
-        workspace.setName("다락방");
-        workspace.setDescription("테스트용 더미 데이터");
-        workspace.setCreatedAt(generateCurrentTimestamp());
-        workspace.setUpdatedAt(generateCurrentTimestamp());
-        workspace.setTokenLimit(50000);
-        workspace.setActive(true);
-
-        return workspace;
+        return WorkspaceEntity
+                .builder()
+                .name("8ee589ef-c7bb-4f2a-a773-630abd0de8c7")
+                .name("다락방")
+                .description("테스트용 더미 데이터")
+                .tokenLimit(50000)
+                .build();
     }
-}
+
+}// end of class
